Provide initial functionality with support for pipelining and transactions
All functions with the exception of multi/0, exec/1, pipe/0, and line/1,
are generated from a list of commands in `lib/exredis_pool.ex`. The functions
generated include the single-shot function to execute the redis command
immediately and the overloaded one for use when the command is part of a
transaction or pipeline.